disable the deep-water feature by default
As said in its commit, it looks odd when pawns have their bodies
hidden in deep water and animals do not. Odyssey introduced extra rendering
for swimming animals, but it is not used for most of them, so if that
gets completed, then it would make sense to enable this by default.